inject only via external scripts
