+++ posthog/plugins/web.py
@@ -32,7 +32,7 @@ def get_transpiled_web_source(id: int, token: str) -> Optional[WebSource]:
     return WebSource(*(list(response)))  # type: ignore
 
 
-def get_transpiled_web_sources(team) -> List[WebSource]:
+def get_decide_web_js_inject(team) -> List[dict]:
     """
     :type team: posthog.models.Team
     """
@@ -45,7 +45,23 @@ def get_transpiled_web_sources(team) -> List[WebSource]:
             plugin__pluginsourcefile__filename="web.ts",
             plugin__pluginsourcefile__status=PluginSourceFile.Status.TRANSPILED,
         )
-        .values_list("id", "plugin__pluginsourcefile__transpiled", "web_token", "plugin__config_schema", "config")
+        .values_list("id", "web_token")
         .all()
     )
-    return [WebSource(*(list(source))) for source in sources]  # type: ignore
+    return [
+        {
+            "id": source[0],
+            "url": f"/web_js/{source[0]}/{source[1]}/",
+        }
+        for source in sources
+    ]
+
+
+def get_web_config_from_schema(config_schema: Optional[List[dict]], config: Optional[dict]):
+    if not config or not config_schema:
+        return {}
+    return {
+        schema_element["key"]: config.get(schema_element["key"], schema_element.get("default", None))
+        for schema_element in config_schema
+        if schema_element.get("web", False) and schema_element.get("key", False)
+    }
